+++ packages/remirror__react-core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @remirror/react-core
 
+## 2.0.18
+
+> 2023-05-24
+
+### Patch Changes
+
+- 60d0a8fd7: Fix the bundled `.d.ts` file for `create-context-hook`.
+- Updated dependencies [60d0a8fd7]
+  - create-context-state@2.0.2
+
 ## 2.0.17
 
 > 2023-04-26